% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{copy_source}
\alias{copy_source}
\title{`copy_source` Copies files from a user-installed package. Intended for
non-standard directories that package developers include under the `inst` 
directory, which installation brings up one level to the top package 
directory and deletes `inst`.}
\usage{
copy_source(
  dest_dir = ".",
  package = "pbdIO",
  kind = "*.sh$",
  source_dir = "slurm"
)
}
\arguments{
\item{dest_dir}{Character string giving the directory where to copy demo shell script.}

\item{package}{Character string giving the package name.}

\item{kind}{Character string giving kind of job manager script, which is also the
directory name for its location in the package source.}

\item{source_dir}{Character string of the package directory containing the file(s) to copy.}
}
\description{
`copy_source` Copies files from a user-installed package. Intended for
non-standard directories that package developers include under the `inst` 
directory, which installation brings up one level to the top package 
directory and deletes `inst`.
}
\details{
Details for \code{pbdIO}:
Running MPI codes on clusters requires batch submission to a job scheduler,
which is best done with a shell script. This function provides a means to 
extract such shell script examples from a package that includes them under 
extra directories that installation makes from `inst`.
}
