% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdIO-package.r
\docType{package}
\name{pbdIO-package}
\alias{pbdIO-package}
\title{pbdIO}
\description{
An interface to parallel input output packages
with a focus on Single Program/Multiple Data ('SPMD')
parallel programming style, which is intended for batch parallel
execution.
}
\references{
Project URL: \url{https://github.com/RBigData/pbdIO}
}
\author{
George Ostrouchov
}
\keyword{package}
