% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm.fread.r
\name{comm.fread}
\alias{comm.fread}
\title{comm.fread}
\usage{
comm.fread(dir, pattern = "*.csv", readers = comm.size(),
  rebalance = TRUE, verbose = 0, checksum = TRUE)
}
\arguments{
\item{dir}{A directory containing the files desired to be read.  The directory
should be accessible to all readers.}

\item{pattern}{The pattern for files desired to be read.}

\item{readers}{The number of readers.}

\item{rebalance}{Logical; should all ranks "balance" the return, meaning each has
roughly the same number of rows?}

\item{verbose}{Determines the verbosity level. Acceptable values are 0, 1, and 2 for
least to most verbosity.}

\item{checksum}{Logical; should numerical variable sums be reported to check input 
before and after rebalance?}
}
\value{
TODO
}
\description{
Given a directory, \code{comm.fread()} reads all csv files contained
in it in parallel with available resources.
}
\examples{
\dontrun{
### Save code in a file "demo.r" and run with 2 processors by
### SHELL> mpiexec -np 2 Rscript demo.r
library(pbdMPI)
library(pbdIO)

path <- "/tmp/read"
comm.print(dir(path))
## [1] "a.csv" "b.csv"

X <- comm.fread(path)

comm.print(X, all.rank=TRUE)
## COMM.RANK = 0
##    a b c
## 1: 1 2 3
## COMM.RANK = 1
##    a b c
## 1: 2 3 4

finalize()
}

}

